Moved the ftt_s_sales_or_investments file to ftt_cpre folder
