+++ SourceCode/ftt_core/ftt_s_sales_or_investments.py
